feat: label extension 분리, reload 방식 변경
UILabel의 extension을 별도의 파일에 분리함
segment가 변경될 때 reload하는 것이 아니라,
새 데이터를 받아오면 insert하는 식으로 변경
